Readability fixes and package name correction
