+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/ToDoListApiApplication.java
@@ -1,0 +1,13 @@
+package school.mindera.toDoListAPI;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ToDoListApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ToDoListApiApplication.class, args);
+	}
+
+}
