+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/controller/UserController.java
@@ -1,0 +1,14 @@
+package school.mindera.toDoListAPI.controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import school.mindera.toDoListAPI.service.UserService;
+
+@RestController
+@RequestMapping("/todo/users")
+public class UserController {
+    private final UserService userService;
+
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
+}
