+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/entities/CommentsEntity.java
@@ -1,0 +1,28 @@
+package school.mindera.toDoListAPI.entities;
+
+import lombok.*;
+import javax.persistence.*;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Entity
+@Table(name = "comments")
+public class CommentsEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer commentId;
+
+    @Column(nullable = false)
+    private String description;
+
+    @ManyToOne
+    @JoinColumn(name = "taskId")
+    private TasksEntity taskId;
+
+    @ManyToOne
+    @JoinColumn(name = "userId")
+    private UsersEntity userId;
+}
