+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/entities/FollowersEntity.java
@@ -1,0 +1,28 @@
+package school.mindera.toDoListAPI.entities;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import school.mindera.toDoListAPI.entities.compositePrimaryKeys.FollowersId;
+import javax.persistence.*;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Entity
+@Table(name = "followers")
+public class FollowersEntity {
+
+    @EmbeddedId
+    private FollowersId id;
+
+    @ManyToOne
+    @MapsId("followingId")
+    private UsersEntity following;
+
+    @ManyToOne
+    @MapsId("userId")
+    private UsersEntity user;
+}
