+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/entities/TagsEntity.java
@@ -1,0 +1,31 @@
+package school.mindera.toDoListAPI.entities;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import javax.persistence.*;
+import java.util.List;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Entity
+@Table(name = "tags")
+public class TagsEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer tagId;
+
+    @Column(nullable = false)
+    private String name;
+
+    @ManyToOne
+    @JoinColumn(name = "userId")
+    private UsersEntity userId;
+
+    @OneToMany(mappedBy = "tag")
+    private List<TaskTagsEntity> tags;
+}
