+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/model/DTOComment.java
@@ -1,0 +1,20 @@
+package school.mindera.toDoListAPI.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class DTOComment {
+
+    private Integer commentId;
+    private String description;
+    private String date;
+    private String userName;
+    private Integer userId;
+    private Integer taskId;
+}
