+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/model/DTOGroup.java
@@ -1,0 +1,15 @@
+package school.mindera.toDoListAPI.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class DTOGroup {
+    private Integer groupId;
+    private String name;
+}
