+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/model/DTOLogin.java
@@ -1,0 +1,12 @@
+package school.mindera.toDoListAPI.model;
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class DTOLogin {
+    private String userName;
+    private String password;
+}
