+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/model/DTONewComment.java
@@ -1,0 +1,19 @@
+package school.mindera.toDoListAPI.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.criteria.CriteriaBuilder;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class DTONewComment {
+    private String description;
+    private String date;
+    private Integer userId;
+    private Integer taskId;
+}
