+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/model/DTONewGroup.java
@@ -1,0 +1,14 @@
+package school.mindera.toDoListAPI.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class DTONewGroup {
+    private String name;
+}
