+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/model/DTONewTag.java
@@ -1,0 +1,15 @@
+package school.mindera.toDoListAPI.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class DTONewTag {
+    String name;
+    String color;
+}
