+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/model/DTORegister.java
@@ -1,0 +1,15 @@
+package school.mindera.toDoListAPI.model;
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class DTORegister {
+private String email;
+private String firstName;
+private String lastName;
+private String userName;
+private  String password;
+}
