+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/model/DTOTag.java
@@ -1,0 +1,16 @@
+package school.mindera.toDoListAPI.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class DTOTag {
+    private Integer tagId;
+    private String name;
+    private String color;
+}
