+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/repository/GroupsRepositoryInterface.java
@@ -1,0 +1,15 @@
+package school.mindera.toDoListAPI.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import school.mindera.toDoListAPI.entities.GroupsEntity;
+
+import java.util.List;
+
+@Repository
+public interface GroupsRepositoryInterface extends JpaRepository<GroupsEntity, Integer> {
+
+    List<GroupsEntity> findByName(String name);
+    List<GroupsEntity> findByUserId(Integer userId);
+
+}
