+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package school.mindera.toDoListAPI.repository;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class UserRepository {
+}
