+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/service/GroupsService.java
@@ -1,0 +1,15 @@
+package school.mindera.toDoListAPI.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import school.mindera.toDoListAPI.repository.GroupsRepositoryInterface;
+
+@Service
+public class GroupsService {
+    @Autowired
+    private final GroupsRepositoryInterface groupsRepository;
+
+    public GroupsService(GroupsRepositoryInterface groupsRepository) {
+        this.groupsRepository = groupsRepository;
+    }
+}
