+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/service/UserService.java
@@ -1,0 +1,15 @@
+package school.mindera.toDoListAPI.service;
+
+import org.springframework.stereotype.Service;
+import school.mindera.toDoListAPI.repository.UserRepository;
+
+@Service
+public class UserService {
+
+    private UserRepository userRepository;
+
+    public UserService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+}
