--- BE/toDoListAPI/src/main/java/shcool/mindera/toDoListAPI/ToDoListApiApplication.java
@@ -1,13 +1,0 @@
-package shcool.mindera.toDoListAPI;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ToDoListApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ToDoListApiApplication.class, args);
-	}
-
-}
