--- BE/toDoListAPI/src/main/java/shcool/mindera/toDoListAPI/entities/GroupsEntity.java
@@ -1,29 +1,0 @@
-package shcool.mindera.toDoListAPI.entities;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.util.List;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@Entity
-@Table(name = "groups")
-public class GroupsEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer groupId;
-    @Column(nullable = false)
-    private String name;
-    @ManyToOne
-    @JoinColumn(name = "userId")
-    private UsersEntity userId;
-
-    @OneToMany(mappedBy = "groupId")
-    private List<TasksEntity> tasks;
-}
