--- BE/toDoListAPI/src/main/java/shcool/mindera/toDoListAPI/model/DTOLogin.java
@@ -1,12 +1,0 @@
-package shcool.mindera.toDoListAPI.model;
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class DTOLogin {
-    private String userName;
-    private String password;
-}
