--- BE/toDoListAPI/src/main/java/shcool/mindera/toDoListAPI/model/DTONewComment.java
@@ -1,19 +1,0 @@
-package shcool.mindera.toDoListAPI.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.criteria.CriteriaBuilder;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class DTONewComment {
-    private String description;
-    private String date;
-    private Integer userId;
-    private Integer taskId;
-}
