--- BE/toDoListAPI/src/main/java/shcool/mindera/toDoListAPI/model/DTONewFollow.java
@@ -1,15 +1,0 @@
-package shcool.mindera.toDoListAPI.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class DTONewFollow {
-    private Integer followerId;
-    private Integer followingId;
-}
