--- BE/toDoListAPI/src/main/java/shcool/mindera/toDoListAPI/model/DTONewGroup.java
@@ -1,14 +1,0 @@
-package shcool.mindera.toDoListAPI.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class DTONewGroup {
-    private String name;
-}
