--- BE/toDoListAPI/src/main/java/shcool/mindera/toDoListAPI/model/DTONewTag.java
@@ -1,15 +1,0 @@
-package shcool.mindera.toDoListAPI.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class DTONewTag {
-    String name;
-    String color;
-}
