--- BE/toDoListAPI/src/main/java/shcool/mindera/toDoListAPI/model/DTONewTask.java
@@ -1,18 +1,0 @@
-package shcool.mindera.toDoListAPI.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.Date;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class DTONewTask {
-    private String title;
-    private String description;
-    private Date finalDate;
-}
