--- BE/toDoListAPI/src/main/java/shcool/mindera/toDoListAPI/model/DTORegister.java
@@ -1,15 +1,0 @@
-package shcool.mindera.toDoListAPI.model;
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class DTORegister {
-private String email;
-private String firstName;
-private String lastName;
-private String userName;
-private  String password;
-}
