--- BE/toDoListAPI/src/main/java/shcool/mindera/toDoListAPI/model/DTOTag.java
@@ -1,16 +1,0 @@
-package shcool.mindera.toDoListAPI.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class DTOTag {
-    private Integer tagId;
-    private String name;
-    private String color;
-}
