--- BE/toDoListAPI/src/main/java/shcool/mindera/toDoListAPI/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package shcool.mindera.toDoListAPI.repository;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class UserRepository {
-}
