--- BE/toDoListAPI/src/main/java/shcool/mindera/toDoListAPI/service/GroupsService.java
@@ -1,15 +1,0 @@
-package shcool.mindera.toDoListAPI.service;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import shcool.mindera.toDoListAPI.repository.GroupsRepositoryInterface;
-
-@Service
-public class GroupsService {
-    @Autowired
-    private final GroupsRepositoryInterface groupsRepository;
-
-    public GroupsService(GroupsRepositoryInterface groupsRepository) {
-        this.groupsRepository = groupsRepository;
-    }
-}
