+++ BE/toDoListAPI/src/test/java/school/mindera/toDoListAPI/ToDoListApiApplicationTests.java
@@ -1,0 +1,13 @@
+package school.mindera.toDoListAPI;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ToDoListApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
