--- BE/toDoListAPI/src/test/java/shcool/mindera/toDoListAPI/ToDoListApiApplicationTests.java
@@ -1,13 +1,0 @@
-package shcool.mindera.toDoListAPI;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ToDoListApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
